Fixed assignment skeleton

+++ Module.cpp
@@ -35,7 +35,7 @@ bool Module::configure (yarp::os::ResourceFinder &rf)
         return false;
     }
 
-    double actuatedDOFs = iCubMainJoints.size();
+    unsigned actuatedDOFs = iCubMainJoints.size();
 
     //create an instance of wbi
     m_robot = new yarpWbi::yarpWholeBodyInterface("computed_torque", wbiProperties);
